REFACTOR: 프리셋 카드 리팩토링
